[refactor] group → team 네이밍 일괄 변경 (패키지) #15
- 패키지이름 group -> team 변경으로 인한 경로 변경
