+++ everyday/src/main/java/com/nyam/everyday/web/team/dto/TeamNoticeDTO.java
@@ -1,4 +1,4 @@
-package com.nyam.everyday.web.groub.dto;
+package com.nyam.everyday.web.team.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -9,20 +9,20 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 /**
- * 그룹 실시간 현황 DTO
+ * 그룹 공지 DTO
  *
  * @author : 이지은
- * @fileName : TeamActivityFeedDTO
+ * @fileName : TeamNoticeDTO
  * @since : 25. 8. 4.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class TeamActivityFeedDTO {
+public class TeamNoticeDTO {
 
     @Schema(name = "", example = "")
-    private Long feedId;
+    private Long teamNoticeId;
 
     @Schema(name = "", example = "")
     private Long teamId;
@@ -31,11 +31,14 @@ public class TeamActivityFeedDTO {
     private Long memberId;
 
     @Schema(name = "", example = "")
-    private String activityType;
+    private String title;
 
     @Schema(name = "", example = "")
-    private String activityContent;
+    private String content;
 
     @Schema(name = "", example = "")
-    private LocalDateTime feedCreatedDate;
+    private LocalDateTime teamNoticeCreatedDate;
+
+    @Schema(name = "", example = "")
+    private LocalDateTime teamNoticeModifiedDate;
 }
