+++ everyday/src/main/java/com/nyam/everyday/web/team/dto/TeamRankingHistoryDTO.java
@@ -1,4 +1,4 @@
-package com.nyam.everyday.web.groub.dto;
+package com.nyam.everyday.web.team.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -9,20 +9,20 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 /**
- * 그룹 공지 DTO
+ * 그룹 내부 랭킹 기록 DTO
  *
  * @author : 이지은
- * @fileName : TeamNoticeDTO
+ * @fileName : TeamRankingHistoryDTO
  * @since : 25. 8. 4.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class TeamNoticeDTO {
+public class TeamRankingHistoryDTO {
 
     @Schema(name = "", example = "")
-    private Long teamNoticeId;
+    private Long teamRankingId;
 
     @Schema(name = "", example = "")
     private Long teamId;
@@ -31,14 +31,14 @@ public class TeamNoticeDTO {
     private Long memberId;
 
     @Schema(name = "", example = "")
-    private String title;
+    private String weekCode;
 
     @Schema(name = "", example = "")
-    private String content;
+    private int point;
 
     @Schema(name = "", example = "")
-    private LocalDateTime teamNoticeCreatedDate;
+    private String field;
 
     @Schema(name = "", example = "")
-    private LocalDateTime teamNoticeModifiedDate;
+    private LocalDateTime teamRankingCreatedDate;
 }
