+++ everyday/src/main/java/com/nyam/everyday/web/team/mapper/TeamNotificationMapper.java
@@ -1,25 +1,25 @@
-package com.nyam.everyday.web.groub.mapper;
+package com.nyam.everyday.web.team.mapper;
 
-import com.nyam.everyday.module.team.entity.TeamNotice;
-import com.nyam.everyday.web.groub.dto.TeamNoticeDTO;
+import com.nyam.everyday.module.team.entity.TeamNotification;
+import com.nyam.everyday.web.team.dto.TeamNotificationDTO;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
 /**
- * 그룹 공지 Builder mapper
+ * 그룹 알림 Builder mapper
  *
  * @author : 이지은
- * @fileName : TeamNoticeMapper
+ * @fileName : TeamNotificationMapper
  * @since : 25. 8. 4.
  */
 @Mapper(componentModel = "spring")
-public interface TeamNoticeMapper {
+public interface TeamNotificationMapper {
 
     @Mapping(source = "Team.TeamId", target = "TeamId")
     @Mapping(source = "member.memberId", target = "memberId")
-    TeamNoticeDTO toDTO(TeamNotice entity);
+    TeamNotificationDTO toDTO(TeamNotification entity);
 
     @Mapping(source = "TeamId", target = "Team.TeamId")
     @Mapping(source = "memberId", target = "member.memberId")
-    TeamNotice toEntity(TeamNoticeDTO dto);
+    TeamNotification toEntity(TeamNotificationDTO dto);
 }
